fix(amd/hip): FP8 MoE coherence on gfx1151 + 35B-A3B strix-hip-real pack
The CUDA->HIP ports of the FP8 kernels read the per-128 block scale as
__nv_bfloat16, but load_fp8_block_scaled_as_fp8weight widens scales to an
FP32 buffer. The 2-byte vs 4-byte stride mismatch plus the bit
reinterpretation corrupted every FP8 weight scale, so Qwen3.6-35B-A3B (MoE)
produced garbage output. The dense 27B was unaffected because it runs the
NVFP4 path. Read the block scales as float to match the loader.

Scale-dtype fixes (block scale __nv_bfloat16 -> float):
- moe_shared_expert_fused_fp8, _batch2, _batch3
- w8a16_gemv, w8a16_gemm, w8a16_gemm_t (incl. transpose_block_scale)

Other fixes:
- moe_fp8_grouped_gemm: replace the no-op stub (only reachable on the dense
  target) with a real WMMA implementation mirroring the gb10 semantics via
  the w8a16_gemm gfx1151 WMMA idiom, reading FP32 scales. Restores coherent
  long-prefill (>64 token) MoE output.
- add widen_block_scale_f32 (referenced by the FP8 SSM path with no source)
- add cuda_fp16.h HIP compat shim
- add the qwen3.6-35b-a3b strix-hip-real kernel pack; vision_encoder ported
  to HIP (64-bit __shfl_down_sync mask, -ffp-contract=off)

Both models serve coherently with MTP on Strix Halo (gfx1151):
27B      ~2.0s TTFT / ~18-20 tok/s decode
35B-A3B  ~260-370ms TTFT / ~32-36 tok/s decode
NVFP4 path is unchanged.

+++ kernels/strix-hip-real/common/w8a16_gemm.cu
@@ -139,7 +139,7 @@ __device__ __forceinline__ void w8a16_wmma_store(
 extern "C" __global__ void w8a16_gemm(
     const __nv_bfloat16* __restrict__ A,            // [M, K] BF16 activations
     const unsigned char* __restrict__ B,             // [N, K] FP8 E4M3
-    const __nv_bfloat16* __restrict__ block_scale,   // [N/128, K/128] BF16
+    const float* __restrict__ block_scale,   // [N/128, K/128] BF16
     __nv_bfloat16* __restrict__ C,                   // [M, N] BF16 output
     unsigned int M,
     unsigned int N,
@@ -190,7 +190,7 @@ extern "C" __global__ void w8a16_gemm(
 
                     unsigned int n_block = gn / FP8_BLOCK;
                     unsigned int k_block = gk / FP8_BLOCK;
-                    float scale = __bfloat162float(block_scale[n_block * k_blocks + k_block]);
+                    float scale = block_scale[n_block * k_blocks + k_block];
 
                     float dequant_val = E4M3_LUT[weight_byte] * scale;
                     smem_B[k][n] = __float2bfloat16(dequant_val);
@@ -212,7 +212,7 @@ extern "C" __global__ void w8a16_gemm(
 /// Each thread handles one FP8 byte → 1 BF16 output.
 extern "C" __global__ void w8a16_dequant(
     const unsigned char* __restrict__ B,             // [N, K] FP8 E4M3
-    const __nv_bfloat16* __restrict__ block_scale,   // [N/128, K/128] BF16
+    const float* __restrict__ block_scale,   // [N/128, K/128] BF16
     __nv_bfloat16* __restrict__ B_bf16,              // [N, K] BF16 output
     unsigned int K,
     unsigned int N
@@ -229,7 +229,7 @@ extern "C" __global__ void w8a16_dequant(
     unsigned int k_blocks = K / FP8_BLOCK;
     unsigned int n_block = n / FP8_BLOCK;
     unsigned int k_block = k / FP8_BLOCK;
-    float scale = __bfloat162float(block_scale[n_block * k_blocks + k_block]);
+    float scale = block_scale[n_block * k_blocks + k_block];
 
     float val = E4M3_LUT[weight_byte] * scale;
     B_bf16[idx] = __float2bfloat16(val);
